requirements.txt
Flask>=3.0.0
tensorflow>=2.20.0
Pillow>=10.0.0
numpy>=2.1.0
Werkzeug>=3.0.0
gunicorn>=22.0.0
